requirements.txt
sqlparse==0.1.18
antlr4-python2-runtime==4.5.2